requirements.txt
langchain
transformers
langchain_huggingface
accelerate
pyyaml
tqdm
matplotlib
seaborn
sentencepiece
protobuf
wandb
isort
black